+++ bootique-kotlin-config/src/main/java/io/bootique/kotlin/config/Extensions.kt
@@ -3,6 +3,7 @@ package io.bootique.kotlin.config
 import io.bootique.BQCoreModule
 import io.bootique.Bootique
 import io.bootique.kotlin.config.modules.KotlinConfigModule
+import io.bootique.kotlin.core.KotlinBootiqueInterface
 
 /**
  * Extensions for config bootstrapping.
@@ -13,4 +14,8 @@ import io.bootique.kotlin.config.modules.KotlinConfigModule
 fun Bootique.withKotlinConfig(): Bootique {
     return this.override(BQCoreModule::class.java)
         .with(KotlinConfigModule::class.java)
+}
+
+fun KotlinBootiqueInterface.withKotlinConfig(): KotlinBootiqueInterface {
+    return this.override(BQCoreModule::class).with(KotlinConfigModule::class)
 }