+++ bootique-kotlin-config/src/main/java/io/bootique/kotlin/config/modules/KotlinConfigModule.kt
@@ -1,14 +1,11 @@
 package io.bootique.kotlin.config.modules
 
-import com.google.inject.Binder
-import com.google.inject.Module
 import io.bootique.config.ConfigurationFactory
-import io.bootique.kotlin.asSingleton
-import io.bootique.kotlin.bind
 import io.bootique.kotlin.config.DefaultKotlinScriptCompiler
 import io.bootique.kotlin.config.KotlinScriptCompiler
 import io.bootique.kotlin.config.KotlinScriptConfigurationFactory
-import io.bootique.kotlin.toClass
+import io.bootique.kotlin.guice.KotlinBinder
+import io.bootique.kotlin.guice.KotlinModule
 
 /**
  * Module for overriding default JsonNodeConfigurationFactory.
@@ -16,9 +13,9 @@ import io.bootique.kotlin.toClass
  * @author Ibragimov Ruslan
  * @since 0.1
  */
-class KotlinConfigModule : Module {
-    override fun configure(binder: Binder) {
-        binder.bind(KotlinScriptCompiler::class).toClass(DefaultKotlinScriptCompiler::class).asSingleton()
-        binder.bind(ConfigurationFactory::class).toClass(KotlinScriptConfigurationFactory::class).asSingleton()
+class KotlinConfigModule : KotlinModule {
+    override fun configure(binder: KotlinBinder) {
+        binder.bind(KotlinScriptCompiler::class).to(DefaultKotlinScriptCompiler::class).asSingleton()
+        binder.bind(ConfigurationFactory::class).to(KotlinScriptConfigurationFactory::class).asSingleton()
     }
 }