+++ RELEASE-NOTES.md
@@ -2,6 +2,8 @@
 
 * Kotlin updated to 1.2.0
 * Dependency on `kotlin-stdlib-jre8` replaced with `kotlin-stdlib-jdk8`
+* New classes: `KotlinModule`, `KotlinBootique`, etc. See documentation for details.
+* Some extensions delete in favor of new classes. See documentation for details.
 
 ## 0.24